wat kleine verbeteringen

+++ www/js/settingsCtrl.js
@@ -1,30 +1,20 @@
 angular.module('calcworks.controllers')
 
 
-.controller('SettingsCtrl', function($scope, $ionicActionSheet, $timeout, sheetService) {
+.controller('SettingsCtrl', function($scope, $ionicPopup, $rootScope, sheetService) {
 
 
     $scope.deleteAllSheets = function() {
-            // show ionic actionSheet to confirm delete operation
-            // show() returns a function to hide the actionSheet
-            var hideSheet = $ionicActionSheet.show({
-                titleText: 'Are you sure that you\'d like to delete all sheets?',
-                cancelText: 'Cancel',
-                destructiveText: 'Delete',
-                cancel: function () {
-                    // do nothing
-                },
-                destructiveButtonClicked: function () {
-                    sheetService.deleteAllSheets();
-                    // hide the confirmation dialog, ik vraag me af of dit wel nodig is
-                    //hideSheet();
-                }
-            });
-
-        $timeout(function() {
-            hideSheet();
-        }, 5000);
-
+        var confirmPopup = $ionicPopup.confirm({
+            title: 'Delete all sheets',
+            template: 'Are you sure you want to delete all sheets?'
+        });
+        confirmPopup.then(function(res) {
+            if (res) {
+                sheetService.deleteAllSheets();
+                $rootScope.$broadcast("allSheetsDeletedEvent", null);
+            }
+        });
     };
 
 });
